AWS_Services/AWS_Project1/worker/requirements.txt
# Worker dependencies
boto3
pandas
pyarrow
s3fs
python-dotenv
